Split sort & print functions in a dedicated file
