+++ main.go
@@ -1,23 +1,9 @@
+// Copyright (c) 2014 Andrea Masi. All rights reserved.
+// Use of this source code is governed by a MIT license
+// that can be found in the LICENSE.txt file.
+
 // RIM - Remote Interfaces Monitor
-
-/* Copyright (c) 2014 Andrea Masi
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE. */
+// Agentless network interfaces monitor for Linux firewalls/servers
 package main
 
 import (
@@ -257,62 +243,6 @@ func parallelizeWorkers(jQueue chan job) {
 	}
 }
 
-func presentSingleResult(r *jobResult) {
-	fmt.Printf("%20.20s%12.12s%8.8s%8.8s\n", "Host", "Interface", "RX-KBps", "TX-KBps")
-	if r.err != nil {
-		fmt.Println(r.host, r.err)
-	} else {
-		// k is remote interface
-		// v is a map with rates
-		for k, v := range r.data {
-			fmt.Printf("%20.20s", r.host)
-			fmt.Printf("%12.12s", k)
-			fmt.Printf("%8d", uint64(v["rx-Bps"]/1024))
-			fmt.Printf("%8d", uint64(v["tx-Bps"]/1024))
-			fmt.Println("")
-		}
-	}
-}
-
-func printHead() {
-	fmt.Printf(
-		"%20s%12s%9s%9s%12s%12s%12s%12s%12s%12s\n",
-		"Host",
-		"Interface",
-		"Rx-Kb/s",
-		"Tx-Kb/s",
-		"Rx-Pckts/s",
-		"Tx-Pckts/s",
-		"Rx-Drp/s",
-		"Tx-Drp/s",
-		"Rx-Err/s",
-		"Tx-Err/s",
-	)
-}
-
-func displayResults(results []interfaceData, noHead bool) {
-	for i, r := range results {
-		if i%40 == 0 && !noHead {
-			printHead()
-		}
-		if r.err != nil {
-			fmt.Fprintln(os.Stderr, "[ERROR]", r.host, r.err)
-		} else {
-			fmt.Printf("%20s", r.host)
-			fmt.Printf("%12s", r.name)
-			fmt.Printf("%9d", uint64(r.rates["rx-Bps"]*8/1024))
-			fmt.Printf("%9d", uint64(r.rates["tx-Bps"]*8/1024))
-			fmt.Printf("%12d", r.rates["rx-pps"])
-			fmt.Printf("%12d", r.rates["tx-pps"])
-			fmt.Printf("%12d", r.rates["rx-dps"])
-			fmt.Printf("%12d", r.rates["tx-dps"])
-			fmt.Printf("%12d", r.rates["rx-eps"])
-			fmt.Printf("%12d", r.rates["tx-eps"])
-			fmt.Println("")
-		}
-	}
-}
-
 func getHostsFromFile(path string) []string {
 	bytes := []byte{}
 	err := errors.New("")
@@ -342,7 +272,7 @@ func main() {
 	versionFlag := flag.Bool("v", false, "Show version and exit")
 	flag.Parse()
 	if *versionFlag {
-		fmt.Println("RIM - Remote Interfaces Monitor v1.0.0")
+		fmt.Println("RIM - Remote Interfaces Monitor v2.0.0-alfa")
 		return
 	}
 	hosts := getHostsFromFile(*hostsFileFlag)
